services/transaction-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
redis==5.2.1
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0
python-jose[cryptography]==3.3.0
alembic==1.14.0
httpx==0.28.1
